chore: rewrite more functions to typescript
