+++ app/api/server/helpers/requestParams.ts
@@ -1,5 +1,5 @@
 import { API } from '../api';
 
-(API as any).helperMethods.set('requestParams', function _requestParams() {
+(API as any).helperMethods.set('requestParams', function _requestParams(this: any) {
 	return ['POST', 'PUT'].includes(this.request.method) ? this.bodyParams : this.queryParams;
 });
